+++ NEWS.md
@@ -5,6 +5,12 @@
   there includes several links that document the options available for each
   browser, along with a few examples.
 - Made fetching the latest release of Selenium Server more reliable.
+- The `error` argument in `wait_for_selenium_available()` now defaults to
+  `TRUE`. This means that the function will throw an error if a Selenium server
+  is not available by default.
+- Added `wait_for_server()`, a version of `wait_for_selenium_available()` that
+  gives more detailed error messages using server created using
+  `selenium_server()`.
 
 # selenium 0.1.3